aider: Added a new command-line flag `-n` or `--net-activity` to enable the display of network events.
# Aider chat conversation:

ASSISTANT: To add a new command-line flag `-n` or `--net-activity` that will enable the display of network events by setting the `ShowNetActivity` bool to true, we need to modify the `init()` function in the `cmd/root.go` file.

Here is the change:

```go
cmd/root.go
<<<<<<< HEAD
func init() {
	// Sub commands removed
}
=======
func init() {
	// Sub commands removed
	RootCmd.PersistentFlags().BoolVarP(&ShowNetActivity, "net-activity", "n", false, "Enable display of network events")
}
>>>>>>> updated
```

This change adds a new persistent flag to the root command. The `BoolVarP` function is used to define a bool flag with specified name, shorthand, default value, and usage string. The address of a bool variable in which to store the value of the flag is provided as the first argument (`&ShowNetActivity`). The flag will be accessible in all subcommands.