aider: Add a command-line flag `-i|--interactive` to enable interactive mode.
# Aider chat conversation:

ASSISTANT: To add a command-line flag `-i|--interactive` that will set the `Interactive` variable to `true`, we need to modify the `init()` function in the `cmd/root.go` file. This function is where command-line flags are defined.

Here is the change:

```go
cmd/root.go
<<<<<<< HEAD
func init() {
	// Sub commands removed
	RootCmd.PersistentFlags().BoolVarP(&ShowNetActivity, "net-activity", "n", false, "Enable display of network events")
}
=======
func init() {
	// Sub commands removed
	RootCmd.PersistentFlags().BoolVarP(&ShowNetActivity, "net-activity", "n", false, "Enable display of network events")
	RootCmd.PersistentFlags().BoolVarP(&Interactive, "interactive", "i", false, "Enable interactive mode")
}
>>>>>>> updated
```

This change adds a new persistent flag `interactive` with a short form `-i`. The flag is a boolean flag, and its default value is `false`. The description of the flag is "Enable interactive mode".