aider: Updated code to only ask the user if they want to be greeted by name if the name variable is not null.
# Aider chat conversation:

USER: please update the code so that if the user dismisses the first input, they are not asked if they want to be greeted by name

ASSISTANT: To implement this change, we need to modify the `assets/dialogs.html` file. We will add a check to see if the `name` variable is not null before asking the user if they want to be greeted by their name.

Here is the *edit block* for the change:

```html
assets/dialogs.html
<<<<<<< HEAD
    <script>
        var name = prompt("Please enter your name", "Harry Potter");
        var config = confirm("Do you want to be greeted by your name?");

        var greeting;
        if (config == true) {
            greeting = "Hello, " + name + "!";
        } else {
            greeting = "Hello, User!";
        }

        alert(greeting);
        document.getElementById("greeting").innerHTML = greeting;
    </script>
=======
    <script>
        var name = prompt("Please enter your name", "Harry Potter");
        var greeting;

        if (name != null) {
            var config = confirm("Do you want to be greeted by your name?");
            if (config == true) {
                greeting = "Hello, " + name + "!";
            } else {
                greeting = "Hello, User!";
            }
        } else {
            greeting = "Hello, User!";
        }

        alert(greeting);
        document.getElementById("greeting").innerHTML = greeting;
    </script>
>>>>>>> updated
```