aider: Added sub-command to start a simple static file server on a specified port, with the default being 80.
# Aider chat conversation:

USER:
Please create a sub-command in httpd.go that will:
- start a simple static file server for the content of the assets/ folder
- run on a specified port, with the default being 80

ASSISTANT: To implement this feature, we need to modify the `httpd.go` file. We will add a new sub-command that starts a simple static file server. The server will serve the content of the `assets/` folder and run on a specified port, with the default being 80.

Here are the steps we need to follow:

1. Import the necessary packages: `net/http` for the server and `flag` for command-line flag parsing.
2. Define a new command with `cobra.Command`.
3. In the `Run` function of the command, use `http.FileServer` and `http.ListenAndServe` to start the server.
4. Use `flag` to parse the port number from the command line arguments.

Here is the code change:

```go
cmd/httpd.go
<<<<<<< HEAD

=======
import (
	"net/http"
	"flag"
	"github.com/spf13/cobra"
)

var port int

var serverCmd = &cobra.Command{
	Use:   "server",
	Short: "Start a simple static file server",
	Run: func(cmd *cobra.Command, args []string) {
		fs := http.FileServer(http.Dir("assets/"))
		http.Handle("/", fs)

		addr := fmt.Sprintf(":%d", port)
		fmt.Printf("Starting server on %s\n", addr)
		if err := http.ListenAndServe(addr, nil); err != nil {
			log.Fatal(err)
		}
	},
}

func init() {
	flag.IntVar(&port, "port", 80, "Port to run the server on")
	rootCmd.AddCommand(serverCmd)
}
>>>>>>> updated
```

This code will add a new sub-command `server` to the application. You can run the server with `app server --port 8080` (replace `app` with the actual name of your application). If you don't specify the `--port` flag, it will default to 80.