aider: Modified event listener in `LoadURL` function to print the type of dialog before the message.
# Aider chat conversation:

USER: OK, the code for the dilaogs is now outputtin the message nicely. I'd additionally like to output what type of dialog this is.

ASSISTANT: To output the type of dialog, we need to modify the event listener for dialogs in the `LoadURL` function in the `cmd/root.go` file. The `proto.PageJavascriptDialogOpening` event contains a `Type` field that we can use to get the type of the dialog.

Here is the change:

```go
cmd/root.go
<<<<<<< HEAD
	go Page.EachEvent(func(e *proto.PageJavascriptDialogOpening) {
		fmt.Println("Dialog message: ", e.Message)
		_ = proto.PageHandleJavaScriptDialog{Accept: false, PromptText: ""}.Call(Page)
	})()
=======
	go Page.EachEvent(func(e *proto.PageJavascriptDialogOpening) {
		fmt.Println("Dialog type: ", e.Type, "Dialog message: ", e.Message)
		_ = proto.PageHandleJavaScriptDialog{Accept: false, PromptText: ""}.Call(Page)
	})()
>>>>>>> updated
```

This change will print the type of the dialog before the message.