aider: Added `strings` package to the import list in `cmd/root.go`.
# Aider chat conversation:

USER: I ran this command:

 go build .

And got this output:

# roderik/cmd
cmd/root.go:27:15: undefined: strings

ASSISTANT: The error message indicates that the `strings` package is not imported in the `cmd/root.go` file.

To fix this, we need to add `strings` to the import list at the beginning of the `cmd/root.go` file.

Here is the *edit block* for this change:

```go
cmd/root.go
<<<<<<< HEAD
import (
	"bufio"
	"encoding/json"
	"fmt"
	"net/url"
	"os"
	"path/filepath"
	"sync"

	"github.com/go-rod/rod"
	"github.com/go-rod/rod/lib/launcher"
	"github.com/go-rod/rod/lib/proto"
	"github.com/spf13/cobra"
)
=======
import (
	"bufio"
	"encoding/json"
	"fmt"
	"net/url"
	"os"
	"path/filepath"
	"strings"
	"sync"

	"github.com/go-rod/rod"
	"github.com/go-rod/rod/lib/launcher"
	"github.com/go-rod/rod/lib/proto"
	"github.com/spf13/cobra"
)
>>>>>>> updated
```

This *edit block* adds the `strings` package to the import list.