aider: Added short command line flag options: -a|--basic-auth and -p|--prt.
# Aider chat conversation:

USER: Please add short command line flag options: -a|--basic-auth and -p|--prt

ASSISTANT: To add short command line flag options, we need to modify the `init` functions in the `cmd/httpd.go` file.

Here are the changes:

```go
cmd/httpd.go
<<<<<<< HEAD
func init() {
	serverCmd.Flags().IntVar(&port, "port", 80, "Port to run the server on")
	serverCmd.Flags().BoolVar(&basicAuth, "basic-auth", false, "Require basic auth")
	RootCmd.AddCommand(serverCmd)
}

func init() {
	flag.IntVar(&port, "port", 80, "Port to run the server on")
	flag.BoolVar(&basicAuth, "basic-auth", false, "Require basic auth")
	RootCmd.AddCommand(serverCmd)
}
=======
func init() {
	serverCmd.Flags().IntVarP(&port, "port", "p", 80, "Port to run the server on")
	serverCmd.Flags().BoolVarP(&basicAuth, "basic-auth", "a", false, "Require basic auth")
	RootCmd.AddCommand(serverCmd)
}

func init() {
	flag.IntVarP(&port, "port", "p", 80, "Port to run the server on")
	flag.BoolVarP(&basicAuth, "basic-auth", "a", false, "Require basic auth")
	RootCmd.AddCommand(serverCmd)
}
>>>>>>> updated
```

In the updated code, I've replaced `IntVar` and `BoolVar` with `IntVarP` and `BoolVarP` respectively. The 'P' at the end of these function names stands for 'Pair', which allows us to specify both a long and a short form of the flag. The short form of the flag is specified as the third argument to these functions.